+++ cmd/cortextool/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## unreleased / master
+## v0.4.0
 
 * [ENHANCEMENT] Loadgen: Allow users to selectively disable query or write loadgen by leaving their respective URL configs empty. #95
 * [FEATURE] Add overrides-exporter to cortextool, which exports Cortex runtime configuration overrides as metrics. #91